+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Application changes:
 Deployment changes:
 
 * Docker launch failed if `HTTPS` environment variable was not passed in and may have failed with a permission denied error.
+* Docker deployments now support environment variables for all application settings.
 * Added a new `branding` environment setting for sites to override templates and provide new assets using a custom Django app.
 * Move mysqlclient dependency from Dockerfile to requirements.in --- pip-compile doesn't have a problem with it anymore.